// src/types/Synergetic/Student/iSynVStudentClass.ts
type iSynVStudentClass = {
  StudentID: number;
  StudentGender: string;
  FileYear: number;
  FileSemester: number;
  ActualSemester: number;
  StudentNameInternal: string;
  StudentNameExternal: string;
  StudentLegalFullName: string;
  StudentSurname: string;
  StudentPreferred: string;
  StudentPreferredFormal: string;
  StudentGiven1: string;
  StudentGiven2: string;
  StudentInitials: string;
  StudentStatus: string;
  StudentLeavingDate: Date | null;
  StudentLeavingReason: string;
  StudentLeavingDestination: string;
  StudentReturningDate: Date | null,
  StudentCampus: string;
  StudentCampusDescription: string;
  StudentYearLevel: number;
  StudentYearLevelDescription: string;
  StudentYearLevelSort: number;
  StudentHouse: string;
  StudentHouseDescription: string;
  StudentHouseHeadofHouse: string;
  StudentHouseCaptain: string;
  StudentForm: string;
  StudentTutor: string;
  StudentBoarder: string;
  StudentBoardingHouse: string;
  StudentBoarderType: string;
  AssessableFlag: boolean;
  ClassColumn: number;
  FileType: string;
  ClassCampus: string;
  ClassCode: string;
  StartDate: Date;
  StopDate: Date | null;
  ClassStatus: string;
  ClassStatusSynergyMeaning: string;
  StudentCalibreMark: boolean;
  AssessmentCode: string;
  ClassDescription: string;
  // ClassNormalYearLevel: null,
  // ClassNormalYearLevel2: null,
  // ClassAdjustmentMean: 0,
  // ClassAdjustmentStandardDeviation: 0,
  ClassFeeCode: string;
  ClassFeeDefaultUnits: number;
  StudentClassFeeUnitsAdjustment: number;
  StudentClassUnitsTotal: number;
  StaffMailName: string;
  SchoolStaffCode: string;
  StaffRegSchoolsBoardCode: string;
  StaffRegistrationNumber: string;
  StaffRegistrationType: string;
  StaffRegistrationExpiryDate: Date | null;
  StaffCampus: string;
  StaffTitle: string;
  StaffPreferredFormal: string;
  StaffPreferred: string;
  StaffNameInternal: string;
  StaffNameExternal: string;
  StaffLegalFullName: string;
  StaffGiven1: string;
  StaffGiven2: string;
  StaffSurname: string;
  StaffInitials: string;
  CurrentSemesterOnlyFlag: boolean;
  StudentEmergencyName: string;
  StudentEmergencyAddress: string;
  StudentEmergencyPhone: string;
  StudentEmergencyPhoneAlt: string;
  StudentEmergencyRelat: string;
  StudentEmergency2Name: string;
  StudentEmergency2Address: string;
  StudentEmergency2Phone: string;
  StudentEmergency2PhoneAlt: string;
  StudentEmergency2Relat: string;
  StudentEmergency3Name: string;
  StudentEmergency3Address: string;
  StudentEmergency3Phone: string;
  StudentEmergency3PhoneAlt: string;
  StudentEmergency3Relat: string;
  StudentStudyAtHomeFlag: boolean;
  StudentIBFlag: boolean;
  StudentFullFeeFlag: boolean;
  StudentTuitionVariationType: string;
  StudentIntendsTertiaryFlag: boolean;
  StudentFamilyPosition: number;
  StudentBirthDate: Date;
  StudentBirthCertificateOnFileFlag: boolean;
  StudentSuburb: string;
  StudentState: string;
  StudentCountryCode: string;
  StudentPostCode: string;
  StudentPhoneActual: string;
  StudentPhoneAlt: string;
  StudentFax: string;
  StudentHomeAddressSameFlag: boolean;
  StudentHomeSuburb: string;
  StudentHomeState: string;
  StudentHomeCountryCode: string;
  StudentHomePostCode: string;
  StudentValidAddressFlag: boolean;
  StudentSilentPhoneFlag: boolean;
  StudentMailFlag: boolean;
  StudentMagazineFlag: boolean;
  StudentSolicitationFlag: boolean;
  StudentRaffleFlag: boolean;
  StudentAddressComma: string;
  StudentAddressFull: string;
  StudentAddressDPIDBarcode: string;
  StudentHomeAddressComma: string;
  StudentHomeAddressFull: string;
  StudentHomeAddressDPIDBarcode: string;
  StudentPhone: string;
  LegalPhotoWebFlag: boolean;
  LegalPhotoPromFlag: boolean;
  LegalPhotoPublicationFlag: boolean;
  StudentStudiesCode: string;
  StudentGovernmentStudentNumber: string;
  ClassLearningAreaCode: string;
  ClassLearningAreaDescription: string;
  ClassLearningAreaReportSortSeq: number;
  ClassLearningAreaActiveFlag: boolean;
  StaffID: number;
  StudentCoursesSeq: number | null;
  StudentSubSchool: string | null;
  InterviewRequestedCode: string;
  InterviewRequestedDescription: string;
  LearningPathway: string;
  StudentGenderSynergyMeaning: string;
  StaffGenderSynergyMeaning: string;
  StudentProfileUrl: string;
}